Changed PostCollection to act as an Enumerable
PostCollection held logic to create Post obects from the wp_export file.
However, in Migrator all it did was pull the array of posts.  This array
later was filtered to non-attachment and valid posts.  I moved the logic
of filtering to the PostCollection and implemented it as a immutable
Enumerable.  Now PostCollection behaves like a primitive array.